+++ src/shared/ui/sidebar/Sidebar.stories.tsx
@@ -12,7 +12,7 @@ import {
 
 import { Sidebar } from './Sidebar'
 
-import { createSidebarItems } from '@/shared/utils/temp/SidebarItem'
+import { createSidebarItems } from '@/shared/utils/sidebarItem/SidebarItem'
 
 const meta: Meta<typeof Sidebar> = {
   title: 'Components/Sidebar',
